requirements_dev.txt
dvc
pandas
ultralytics
mlflow
pyyaml
gdown
-e .